Corrects the description of the 'trace' level of logging
Fixes #463

The 'trace' level of logging is the least severe level of logging in SwiftLog
and should be used only when tracing the execution flow of a program during
debugging.

This description of the 'trace' level of logging, is verbatim the description
of the 'trace' level in the SwiftLog documentation at
https://github.com/apple/swift-log/blob/74d7b91ceebc85daf387ebb206003f78813f71aa/Sources/Logging/Logging.swift#L327
and I will file a bug report and PR there in the upstream project.

Signed-off-by: zach wick <zach@zachwick.com>