Update hello-world.md (#508)
The previous instructions suggested that you could continue with "Build & Run" as soon as Xcode opens, but this is not correct and the Play button is at this point greyed out as unavailable, as is selecting "Build and Run" from the menu, as dependencies are still being downloaded. This could lead users inexperienced with Xcode to think they have done something wrong when they just need to wait a minute or so. This seems likely the cause of at least one comment I saw on Vapor discussion somewhere today when searching for a new user guide (sorry, location lost) where the new user stated that Step 2 of these instructions did not work.

This change also correctly identifies the expected output will be in the Xcode console, and not in the "terminal".