Update references to `app.swift` (#529)
* Update references to `app.swift`

When performing `vapor new hello -n` command as part of [Getting Started](https://docs.vapor.codes/4.0/hello-world/), an `app.swift` file is not generated. The code that creates and runs the `app` lives in `main.swift`. This updates the documentation to reflect this.

* Moved reference to `main.swift` file under "Run"

Created a `### Run` heading, moving the information about `main.swift` under there to match the folder structure on disk.