Fix an invalid example password in Authentication (#533)
The example in authentication -> Fluent first establishes code that checks wether a user’s password is at least 8 characters long, and then uses a 6 character password in the example. This is not a huge problem, but it might present a stumbling block for readers who didn’t pay close attention to the code.