Make custom Leaf tag examples complete
As reported in [vapor/vapor#2942](https://github.com/vapor/vapor/issues/2942), UnsafeUnescapedLeafTag is going unnoticed.

To bring more attention to it, I moved the tip to the top section in all languages, updated the HelloTag to use the protocol (since it was rendering garbage since the LeafKit update).
Plus, all the code samples are now directly copiable, as they have a common structure and indentation.