Add a tip about model middleware to authentication example (#863)
The example we use for `ModelAuthenticatable` is a `User` identified by
an email address. Email addresses are case-insensitive, while the
`ModelAuthenticatable` that ships with `Fluent` is case sensitive. This
means that an application that follows the example naively would
consider `SomeUser@example.com` and `someuser@example.com` to be two
different users. This is easily addressed with model middleware that
converts the email address to lowercase before saving a `User`, but that
example is elsewhere in the documentation.

This change simply calls out this possibility in a tip section and links
to the model middleware section of the docs.